feat: finished B A S I C settings management in svelte ui
